RetrivalAndGenerator/requirements-base.txt
sentence_transformers
llama_cpp_python
torch
scikit-learn
numpy
pymupdf
langchain
accelerate
